Fixing IdLabelConfig import issues
