+++ TestProjects/PerceptionURP/Assets/Scenes/SampleScene.unity
@@ -440,7 +440,7 @@ Light:
   m_BounceIntensity: 1
   m_ColorTemperature: 6570
   m_UseColorTemperature: 0
-  m_BoundingSphereOverride: {x: 1e-45, y: 9e-44, z: 9e-44, w: 0}
+  m_BoundingSphereOverride: {x: 0, y: 1e-45, z: 1.8643e-41, w: 5.9e-44}
   m_UseBoundingSphereOverride: 0
   m_ShadowRadius: 0
   m_ShadowAngle: 0
@@ -615,7 +615,7 @@ MonoBehaviour:
       data:
         enabled: 1
         objectCountMetricId: 51DA3C27-369D-4929-AEA6-D01614635CE2
-        m_IdLabelConfig: {fileID: 11400000, guid: cedcacfb1d9beb34fbbb231166c472fe,
+        m_LabelConfig: {fileID: 11400000, guid: cedcacfb1d9beb34fbbb231166c472fe,
           type: 2}
     00000001:
       type: {class: BoundingBox2DLabeler, ns: UnityEngine.Perception.GroundTruth,
@@ -640,6 +640,7 @@ MonoBehaviour:
         enabled: 1
         annotationId: 12F94D8D-5425-4DEB-9B21-5E53AD957D66
         labelConfig: {fileID: 11400000, guid: c140c5aa05dd09e4fadaa26de31b1f39, type: 2}
+        m_TargetTextureOverride: {fileID: 0}
 --- !u!114 &963194231
 MonoBehaviour:
   m_ObjectHideFlags: 0
